Create boxes with Javascript
Create a function to calculate the number of total boxes for the grid number provided as an argument.
With a loop create boxes, set the width and height of a box according to the total box count and append each box to the container div.
Also set a class attribute for styling.
Set grid container as a flexbox in 960px height and 960px width.
Allow flexbox to wrap after desired number of columns reached in a row with the min-width parameter set in JS.